requirements.txt
feast[postgres,redis]==0.40.1
mlflow==2.16.2
kafka-python-ng==2.2.3
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0
requests>=2.31.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
boto3>=1.28.0
flask>=3.0.0
joblib>=1.3.0
